+++ e2e-tests/tests/sendonly_recvonly.test.ts
@@ -1,3 +1,4 @@
+import { randomUUID } from 'node:crypto'
 import { expect, test } from '@playwright/test'
 
 test('sendonly/recvonly pages', async ({ browser }) => {
@@ -8,6 +9,12 @@ test('sendonly/recvonly pages', async ({ browser }) => {
   // それぞれのページに対して操作を行う
   await sendonly.goto('http://localhost:9000/sendonly/')
   await recvonly.goto('http://localhost:9000/recvonly/')
+
+  const channelName = randomUUID()
+
+  // チャンネル名を設定
+  await sendonly.fill('#channel-name', channelName)
+  await recvonly.fill('#channel-name', channelName)
 
   // SDK バージョンの表示
   await sendonly.waitForSelector('#sdk-version')
@@ -21,23 +28,23 @@ test('sendonly/recvonly pages', async ({ browser }) => {
   await sendonly.click('#connect')
   await recvonly.click('#connect')
 
-  // #sendrecv1-connection-id 要素が存在し、その内容が空でないことを確認するまで待つ
+  // sendonly の #connection-id 要素が存在し、その内容が空でないことを確認するまで待つ
   await sendonly.waitForSelector('#connection-id:not(:empty)')
 
-  // #sendonly-connection-id 要素の内容を取得
+  // sendonly の #connection-id 要素の内容を取得
   const sendonlyConnectionId = await sendonly.$eval('#connection-id', (el) => el.textContent)
   console.log(`sendonly connectionId=${sendonlyConnectionId}`)
 
-  // #sendrecv1-connection-id 要素が存在し、その内容が空でないことを確認するまで待つ
+  // recvonly の #connection-id 要素が存在し、その内容が空でないことを確認するまで待つ
   await recvonly.waitForSelector('#connection-id:not(:empty)')
 
-  // #sendrecv1-connection-id 要素の内容を取得
+  // recvonly の #connection-id 要素の内容を取得
   const recvonlyConnectionId = await recvonly.$eval('#connection-id', (el) => el.textContent)
   console.log(`recvonly connectionId=${recvonlyConnectionId}`)
 
   // レース対策
-  await sendonly.waitForTimeout(5000)
-  await recvonly.waitForTimeout(5000)
+  await sendonly.waitForTimeout(3000)
+  await recvonly.waitForTimeout(3000)
 
   // 'Get Stats' ボタンをクリックして統計情報を取得
   await sendonly.click('#get-stats')