+++ playwright.config.mjs
@@ -16,13 +16,52 @@ export default defineConfig({
         '--use-fake-ui-for-media-stream',
         '--use-fake-device-for-media-stream',
         // "--use-file-for-fake-video-capture=/app/sample.mjpeg",
+
+        '--enable-features=WebRtcAllowH265Send,WebRtcAllowH265Receive',
       ],
     },
   },
   projects: [
+    // Chromium
     {
-      name: 'chromium',
+      name: 'Chromium',
       use: { ...devices['Desktop Chrome'] },
+    },
+
+    // Chrome
+    {
+      name: 'Google Chrome',
+      use: { ...devices['Desktop Chrome'], channel: 'chrome' },
+    },
+    {
+      name: 'Google Chrome Beta',
+      use: { ...devices['Desktop Chrome'], channel: 'chrome-beta' },
+    },
+    {
+      name: 'Google Chrome Dev',
+      use: { ...devices['Desktop Chrome'], channel: 'chrome-dev' },
+    },
+    {
+      name: 'Google Chrome Canary',
+      use: { ...devices['Desktop Chrome'], channel: 'chrome-canary' },
+    },
+
+    // Edge
+    {
+      name: 'Microsoft Edge',
+      use: { ...devices['Desktop Edge'], channel: 'msedge' },
+    },
+    {
+      name: 'Microsoft Edge Beta',
+      use: { ...devices['Desktop Edge'], channel: 'msedge-beta' },
+    },
+    {
+      name: 'Microsoft Edge Dev',
+      use: { ...devices['Desktop Edge'], channel: 'msedge-dev' },
+    },
+    {
+      name: 'Microsoft Edge Canary',
+      use: { ...devices['Desktop Edge'], channel: 'msedge-canary' },
     },
 
     // {